Create audio class and inheritance scheme for sound and music
