+++ simian/audio/audio_manager.py
@@ -1,0 +1,11 @@
+import pygame
+
+class SoundManager:
+    def __init__(self):
+        pygame.mixer.init()
+
+    def pause_all(self):
+        pygame.mixer.pause()
+
+    def unpause_all(self):
+        pygame.mixer.unpause()
