--- simian/audio/sound_manager.py
@@ -1,8 +1,0 @@
-import pygame
-
-class SoundManager:
-        def pause_all(self):
-            pygame.mixer.pause()
-
-        def unpause_all(self):
-            pygame.mixer.unpause()
